Move Album and AddCoverArtToAlbumRunnable to new package com.restmash.album
