+++ src/main/java/com/restmash/app/Artist.java
@@ -1,5 +1,6 @@
 package com.restmash.app;
 
+import com.restmash.album.Album;
 import com.restmash.musicbrainz.MusicBrainzContent;
 import com.restmash.wikipedia.WikipediaContent;
